Isolate sighash to a module
