+++ common/src/chain/transaction/signature/inputsig/authorize_pubkeyhash_spend.rs
@@ -80,7 +80,7 @@ mod test {
     use super::*;
     use crate::chain::signature::tests::utils::generate_inputs_utxos;
     use crate::chain::{
-        signature::{inputsig::StandardInputSignature, signature_hash},
+        signature::{inputsig::StandardInputSignature, sighash::signature_hash},
         transaction::signature::tests::utils::{generate_unsigned_tx, sig_hash_types},
         Destination,
     };
